+++ examples/resources/upcloud_managed_object_storage_user_access_key/resource.tf
@@ -12,4 +12,5 @@ resource "upcloud_managed_object_storage_user" "this" {
 resource "upcloud_managed_object_storage_user_access_key" "this" {
   username     = upcloud_managed_object_storage_user.this.username
   service_uuid = upcloud_managed_object_storage.this.id
+  status       = "Active"
 }